refactor : 에러 메시지 상수 처리

+++ src/main/java/hyper/run/domain/game/dto/response/GameHistoryResponse.java
@@ -29,7 +29,7 @@ public class GameHistoryResponse {
 
     private int participatedCount;
 
-    private int myPrize;
+    private double myPrize;
 
     public static GameHistoryResponse toResponse(Game game , GameHistory gameHistory){
         return GameHistoryResponse.builder()
